+++ R/plot_corr.R
@@ -89,6 +89,13 @@ plot_corr <- function(coin, dset, iCodes = NULL, Levels = 1, ..., cortype = "pea
     iCodes <- rev(iCodes)
   }
 
+  if(withparent == "family"){
+    # in this case we don't care about the second entry and copy from 1
+    # to avoid any issues
+    Levels[2] <- Levels[1]
+    iCodes[[2]] <- iCodes[[1]]
+  }
+
   crtable <- get_corr(coin, dset = dset, iCodes = iCodes, Levels = Levels,
                       ... = ..., cortype = cortype, pval = pval, withparent = withparent,
                       grouplev = grouplev, make_long = TRUE, use_directions = use_directions)
@@ -171,10 +178,15 @@ plot_corr <- function(coin, dset, iCodes = NULL, Levels = 1, ..., cortype = "pea
     crtable$Flag[(crtable$Correlation <= weakthresh)] <- "Weak"
     crtable$Flag[(crtable$Correlation <= negthresh)] <- "Negative"
 
+    # make factors
+    # note: moved from inside ggplot call below to hopefully help ggplotly tooltip
+    crtable$Var1 <- factor(crtable$Var1, levels = ord1)
+    crtable$Var2 <- factor(crtable$Var2, levels = ord2)
+
     # heatmap plot
     plt <- ggplot2::ggplot(data = crtable,
-                           ggplot2::aes(x = factor(.data$Var1, levels = ord1),
-                                        y = factor(.data$Var2, levels = ord2),
+                           ggplot2::aes(x = .data$Var1,
+                                        y = .data$Var2,
                                         fill = .data$Flag,
                                         label = .data$Correlation)) +
       ggplot2::geom_tile(colour = "white") +
@@ -198,12 +210,19 @@ plot_corr <- function(coin, dset, iCodes = NULL, Levels = 1, ..., cortype = "pea
     )
   } else {
 
+    # note: moved from inside ggplot call below to hopefully help ggplotly tooltip
+    crtable$Var1 <- factor(crtable$Var1, levels = ord1)
+    crtable$Var2 <- factor(crtable$Var2, levels = ord2)
+
+    # create duplicate column to be able to turn off tooltip with ggplotly
+    crtable$Correlation2 <- crtable$Correlation
+
     # heatmap plot
     plt <- ggplot2::ggplot(data = crtable,
-                           ggplot2::aes(x = factor(.data$Var1, levels = ord1),
-                                        y = factor(.data$Var2, levels = ord2),
+                           ggplot2::aes(x = .data$Var1,
+                                        y = .data$Var2,
                                         fill = .data$Correlation,
-                                        label = .data$Correlation)) +
+                                        label = .data$Correlation2)) +
       ggplot2::geom_tile(colour = "white") +
       ggplot2::labs(x = NULL, y = NULL, fill = "Correlation") +
       ggplot2::theme_classic() +