+++ NEWS.md
@@ -1,4 +1,22 @@
-# COINr 1.1.1
+# COINr 1.1.7
+
+I have skipped a number of version number iterations. Here's what's changed since last time:
+
+- Add some helper functions to convert codes to names
+- Add flip axes to bar chart
+- Add option to skip SA address checks
+- Option to use raw data in results tables
+- Small bug fixes
+- Catch more errors on coin construction
+
+
+# COINr 1.1.2.9000
+
+Dev version of COINr until next CRAN submission!
+
+* `Treat()` now catches errors - data frame/coin/purse methods 
+
+# COINr 1.1.2
 
 Lots of small fixes and feature extensions, including
 
@@ -8,6 +26,8 @@ Lots of small fixes and feature extensions, including
 * Reordering for correlation plots
 * Better input checks
 * Extra stats
+* Bi-direction log transformation for negative skewed outlier treatment
+* Subgroups of units for bar charts
 
 # COINr 1.1